Fake an FTS service so we can test a full FTS backend with no external services or files. Currently only fakes the 'submit' and 'status' commands, and returns immediate success for all transfers.
 To use it, set PHEDEX_GLITE_WRAPPER in your environment, pointing to this
script. The GliteAsync backend will pick it up automatically.

 You also need to ensure that /tmp/phedex exists and is writeable by this
script, or edit the script and set $cache to some suitable path. This is
needed so that the script can cache the relationship between the fake jobids it
generates and the copyjob file that was submitted with the job, it does this
using a soft-link in this cache directory. It needs that association so it can
give the correct filenames in the status commands which will follow submission.