- Remove a bunch of long-obsolete code related to picking workers.
- Change the 'exit' to 'POE::Kernel->alarm_remove_all' in doStop. This will
kill the POE session, after which the POE kernel will naturally end if this
is the only session running, and the agent will then exit. If there are
several agents in the same process, the rest will continue uninterrupted.

- Do not unlink the STOPFLAG on exit. This leaves the file in place, so the
watchdog can know that the agent was stopped deliberately. The watchdog will
not then restart the agent. If the agent is restarted in the usual manner
(via Utilities/Master) the agent will remove the old stop-file as it starts.

- Add a call to 'Notify' in the agent startup, so the watchdog can pick up
the relation between agent->ME() and the agent-label from the configuration
file. These are not the same, which is a little unfortunate.