PHEDEX::Core::Config::Factory is now a fully-fledged watchdog agent, and can start agents in the same process or externally, using the JobManager to call Utilities/Master in the appropriate manner.
It will 'ping' the agents (with a kill => 0) every cycle, and will restart
any that are down. It will not start agents that have a stop-file in place,
I assume these were stopped deliberately and will be restarted manually when
needed.

If PHEDEX_NOTIFICATION_PORT is set in the environment of any agent, it will
automatically send its Warnings and Alerts to this port by UDP. The Watchdog
agent (implemented in this module) will listen on that port and will add those
messages to its own logfile. That makes this a one-stop place to look for all
agent-alerts.

Agent statistics are also gathered here, the process-level stats on CPU and
memory use etc. Eventually, this agent could spot and kill processes that are
using too much memory.

The watchdog agent will keep track of how long it has been since it received
any message from the agents it is following. If it goes too long in silence,
it will kill and restart the agent. It kills via a succession of signals
(HUP, TERM, KILL) at respectable intervals.