add a PHEDEX.DataTable.ContextMenu class to the same file. This is used by the PHEDEX.Component.ContextMenu decorator to augment the module with the support features needed for a context-menu.
This means I keep the code for datatable-contextmenus in a separate part
of the namespace and create them only when I need them. Much cleaner than
having them all through the file.

The mechanism is generic. The decorator receives the object it must
decorate, and expects a 'type' variable to denote the class of object
('DataTable' or 'LinkView'). It then looks for a sub-type of
PHEDEX[obj.type] that has the same name as the decorator, and uses it
to augment the object.