lots of changes: - sorted out decorator construction cycle and notifications to navigator.   decorators now interact correctly. - addToHistory simplified, uses callbacks into the decorators without   knowing how many or what they are. Decorators notify the navigator that   they can provide state information, but only if they can. Not all do... - addToHistory updated by notification from the sandbox when something has   changed. Specifically, change of instance, target, or widget will trigger   a change. Change of target-type does not trigger a change directly,   because change of target-type always causes a change of widget, so that   is enough to get the history updated. - setState and getCurrentState also using the callbacks into the decorators.   It's almost working, needs more debugging. - removed much speculative code that doesn't seem like it will ever be   needed (and which wasn't used)
 The general direction this is going in is that the navigator will not own
or manage global state directly. Instead, it will manage a number of
decorators, which each own a piece of the state. The navigator will know
how to calculate global state from the state of the decorators (without
knowing anything about their internals).

 The navigator then simply calculates global state and records it via the
history-manager as the application evolves. The decorators can steer the
app based on what the navigator tells them or on what the user does. It
should be possible to add more stateful decorators (or re-implement them
differently) without having to touch the navigator. Similarly, removing
decorators should not break anything (e.g, remove the Instance selector!)