Implementation of methods_allowed() mechanism:
API may define optional methods_allowed().
It returns either a string or a reference to an array of strings.
Each string is a method name, such as 'POST', 'GET', or anything.

If methods_allowed() is defined, the value of request_method() is
compared against the value(s) of methods_allowed(). If the comparison
fails, the call is rejected.

Absence of methods_allowed() means no restriction at all.