rework the flow to eliminate recent problems. Agent now has one controlling loop for fetching work, and another (triggered by the arrival of new work) to execute the tests.
Tests that fail are retried up to 10 times, with one-minute delay from the
previous attempt, and at lower priority than any currently-queued work (i.e.
lower than work in the current cycle, not lower than all work in the DB).

This allows tests to be retried in a reasonable manner, without blocking
the agent. If a number of tests are failing repeatedly, they could still
block the agent until they have been retried to death, but that's OK. If
lots of tests are failing then something more serious is probably wrong.

N.B. 'test failing' refers, in that paragraph, to the failure of the
infrastructure, not the presence or absence of data.