allow the watchdog to kill agents pre-emptively, based on limits set in its configuration file. The watchdog will then restart the agent, as though it had crashed. This will allow us to have some breathing-space in the event of bugs that cause agents to leak memory, for example.
 The config-file syntax is:

 --limit regex,key,value
 '--limit' may be repeated as often as required to impose more constraints

 'regex' is matched against the agent label, and is a regular expression as
its name implies. You can be precise, giving the full name of the agent, or
use wildcards. N.B. in the case that there are multiple agents with names
that are similar, you will need to exercise caution. E.g, if you have a
'download' and a 'download-mss' agent, imposing a limit with

 --limit download,...,...

 will limit both of them, because it matches both as a regex. Use 'download$'
to limit only the agent with exactly that name.

 'key' is one of the standard monitored resources for CPU or memory use. Valid
values are: 'rss','vsize','stime','utime', which are the RSS, VSize, system
and user CPU times since the agent began. You can also specify the same values
with a prefix of 'd' ('drss') to limit the change (delta) of resource use in
the last hour.

 'val' is the numerical upper limit for the use of that resource before the
agent is killed.

 Each key-value can be constrained for an agent, or different values for
different agents, as you wish. So, to limit the exp-pfn agent to an RSS of <
100 MB, and a total (user) CPU of less than 3 hours, use the following:

 --limit exp-pfn,rss,100
 --limit exp-pfn,utime,10800