connectAgent calls checkNodes wether MYNODE is set or not
checkNodes checks if MYNODE is set before checking the value, and checks the
other options it is responsible for regardless.

expandNodes now correctly restores the state of MYNODE if there are errors.
Rather than just setting it to 'undef', the value of MYNODE is recorded at
the beginning of the block and restored after an eval around the sensitive
bit.