add a 'replaceEvent' function, which allows an object to intervene in the core workflow. The function takes three arguments: 1) the name of the event to be replaced 2) the name of the event to replace it with, 3) an flag to indicate if the mapping is one-way (set to 'true')
For example: Calling
 _sbx.replaceEvent('original','replacement');

 has the effect that when any object calls
 _sbx.notify('original',...);

 the actual notification is for a 'replacement' event, with the same
arguments. The fiendishly clever piece of software that listens for the
'replacement' event can then do what it likes, and eventually it can call

 _sbx.notify('replacement',...);
 which the sandbox dutifully reverse-maps it to notify an 'original' event.
The application then continues as normal!

This is used by the navigator to intercept 'CreateModule' events, so it can
update its own state first, and use that to drive the eventual creation.

The mapping is two-way as shown unless the third argument is set to 'true',
in which case a 'replacement' event will not be mapped to an 'original'. Use
this if ever you need more complex (3-way) mapping (which requires 3 calls
to this function)