sorting now applied to all open branches of the same level with a common grandparent. I.e. if you open two top-level branches and sort one of them, the other will also be sorted by the same field.
N.B. sorting is applied only to first-cousins, i.e. those branches with a
common grandparent. If the relationship is deeper (great-grandparent), then
the sorting is not applied to the distant relative, even if they are
highlighted as though it were.

N.N.B. That caveat applies only to already-loaded branches at the time of
sorting. If you open a branch 3-levels down, sort, then fetch data for
other branches at that depth, they will be sorted correctly. It is only
branches that already exist and are too distantly related that do not
get sorted.

I can fix this if someone screams, otherwise it may not be worth the
effort.