add some profiling support. This will build a YUI datatable with information about the calls that take the most total CPU.
Only calls that contribute over 1% of the total profiled time are shown, to
limit clutter.

The profiler assumes that a div with an id 'phedex-profiler' exists somewhere
in the DOM, without which it will silently die.

Profile statistics are updated every 10 seconds.

The profiler also logs detailed information with 'profile/debug' class. If
you are using the proxy, you can capture this in a logfile for further
analysis.