add context-menu entry to view module data as JSON in a popup window. This needs a deep-copy of the data, converting arrays to objects. This is needed because some APIs return arrays with named-keys, not numeric, so the JSON stringifier chokes on them. This appears to be some standard feature of the language, rather than a bug.
For modules that use a _processData, there is no problem, as long as the
table-rows are declared to be objects ('{}') instead of arrays ('[]'). This
is true for all except the 'nodes' api, which does not have an _processData
method. Because the API on the server uses an array, the JSON formatter
(JSON::XS perl module) spits out the 'wrong' type of array for this API, so
we are forced to guard against it.

In any case, it's not nice to have to get the server-side API correct to
cope with limitations of a client-side functionality, especially when there
is no clear difference between the two options, so we don't lose much by
having the formatting done again.