Add procedures proc_add_node and proc_delete_node. I bet you can't guess what they do?
They both contain the logic for creating or deleting the partitions in the
t_xfer_* tables, as well as the inserts/deletions in t_adm_node. This means
that by granting execute privilege on these procedures they can then be
run by someone who does _not_ have the alter-table privilege. Until now,
that has been required to explicitly add or delete a partition.

N.B. I tried doing this with triggers on insert/delete into t_adm_node
instead, but that was not reliable. You cannot trigger on the 'commit',
so if the user rolled back the transaction instead, you would end up
with an inconsistent state. Worst case, the partitions would be deleted
but the node would remain, and could still be used.

NodeNew and NodeRemove have been updated to use these procedures.