Replace an 'update where (select ...)' statement with a 'merge into' statement for updating the priority in t_xfer_request.
The original construct with 'update where' did not work with restricted
column access. It turns out you need full update access to the tables in
the 'select' clause, even if you're not updating all the columns.

Circumvent this problem by using a 'merge into' where the 'select' clause
picks the rowids of the rows to be updated and the values to update them
to, then the 'merge into' matches on rowid and does the update.

This needs proper testing before release, of course!